To allow SQLCoverage tests to run with either HSqlDB (as always) or PostgreSQL (new!) to be the database against which VoltDB results are compared, split old HsqlBackend into abstract base class GenericBackend with sub-classes HsqlBackend and PostgreSQLBackend, each handling the specified database.
